+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
   </head>
   <body>
+    injection failure!
   </body>
-  <script src='./main.js'></script>
+  <script type='module' src='./main.js'></script>
 </html>
